[#9067] MCVE for bug
This MCVE doesn't follow the standard pattern. Instead it uses
Testcontainers in `MCVETest` as well as running the `XMLGenerator` and
`JavaGenerator` programmatically.

Using Testcontainers makes the test very concise and also very easy to
use (no need to manually set up a database of the correct vendor and
version). And for simple cases like this one it also allows skipping
Flyway.

Running the generator programmatically is also just an experiment. It
has the advantage that the code is easier to debug (just debug the JUnit
test straight from Eclipse).